add pojectshop

+++ src/main/java/edu/poly/shop/dao/CategoryDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import edu.poly.shop.model.Category;
 
-public interface CategoryDAO extends JpaRepository<Category, Integer> {
+public interface CategoryDAO extends JpaRepository<Category, Long> {
 
 }
